moved the cmd module to a folder and cmd::Error to it's own module within cmd
