+++ src/cmd/cmd.rs
@@ -1,3 +1,4 @@
+use super::error::Error;
 use std::convert::TryFrom;
 use std::ffi::OsStr;
 use std::io::{self, Write};
@@ -36,35 +37,6 @@ pub enum Op {
 
 #[derive(Debug)]
 pub struct LineIter<'a>(SplitWhitespace<'a>);
-
-// TODO: implement std::error::Error for Error
-#[derive(Debug)]
-pub enum Error {
-    EmptyLine,
-    Io(io::Error),
-}
-
-impl std::fmt::Display for Error {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        match self {
-            Error::Io(e) => write!(f, "{}", e),
-            Error::EmptyLine => Ok(()),
-        }
-    }
-}
-
-impl std::error::Error for Error {
-    fn cause(&self) -> Option<&dyn std::error::Error> {
-        match self {
-            Error::EmptyLine => None,
-            Error::Io(e) => Some(e),
-        }
-    }
-
-    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
-        None
-    }
-}
 
 impl<'a> TryFrom<&'a str> for Expression<'a> {
     type Error = Error;
