+++ src/cmd/mod.rs
@@ -1,0 +1,5 @@
+mod cmd;
+mod error;
+
+pub use cmd::Expression;
+pub use error::Error;
